+++ src/main/java/org/plumelib/util/package-info.java
@@ -91,6 +91,9 @@
  *   <dt>{@link org.plumelib.util.ClassDeterministic ClassDeterministic}
  *   <dd>Deterministic versions of {@code java.lang.Class} methods, which return arrays in sorted
  *       order.
+ *   <dt>{@link org.plumelib.util.ClassDeterministic UniqueId}
+ *   <dd>An interface for objects that have a unique ID. If you are tempted to print the value of
+ *       {@code System.identityHashCode()}, consider using this instead.
  *   <dt>{@link org.plumelib.util.ImmutableTypes ImmutableTypes}
  *   <dd>Indicates which types in the JDK are immutable.
  * </dl>